Split into util and tests file
